move /auth/edit to /private/edit upgrade npm repository order validations signup
